--- src/test/java/com/example/surveyapp/domain/admin/service/BlackListServiceTest.java
@@ -1,146 +1,0 @@
-package com.example.surveyapp.domain.admin.service;
-
-import com.example.surveyapp.domain.admin.domain.model.BlackList;
-import com.example.surveyapp.domain.admin.domain.repository.BlackListRepository;
-import com.example.surveyapp.domain.user.domain.model.User;
-import com.example.surveyapp.domain.user.domain.repository.UserRepository;
-import com.example.surveyapp.global.response.ApiResponse;
-import com.example.surveyapp.global.response.exception.CustomException;
-import com.example.surveyapp.global.response.exception.ErrorCode;
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.http.HttpStatus;
-
-
-import java.util.Optional;
-
-import static org.mockito.Mockito.when;
-
-@ExtendWith(MockitoExtension.class)
-class BlackListServiceTest {
-
-
-    @Mock
-    private BlackListRepository blackListRepository;
-
-    @Mock
-    private UserRepository userRepository;
-
-    @Mock
-    private User user;
-
-    @Mock
-    private BlackList blackList;
-
-    @InjectMocks
-    private BlackListService blackListService;
-
-    @Test
-    void 블랙리스트에_등록한다() {
-
-        // given
-        Long userId = 1L;
-        when(userRepository.findById(userId)).thenReturn(Optional.of(user));
-
-        // when
-        User result = blackListService.addBlackList(userId);
-
-        // then
-        Assertions.assertThat(user).isEqualTo(result);
-
-    }
-
-    @Test
-    void 블랙리스트_등록을_실패한다1() {
-
-        // given
-        Long userId = 1L;
-
-        // when & then
-        try {
-            blackListService.addBlackList(userId);
-        } catch (CustomException e) {
-            Assertions.assertThatExceptionOfType(CustomException.class)
-                    .isThrownBy(() -> { throw new CustomException(e.getErrorCode()); })
-                    .withMessageContaining("사용자를 찾을 수 없습니다.");
-        }
-
-    }
-
-    @Test
-    void 블랙리스트_등록을_실패한다2() {
-
-        // given
-        Long userId = 1L;
-        when(userRepository.findById(userId)).thenReturn(Optional.of(user));
-//        when(blackListRepository.findByUserId(user)).thenReturn(Optional.of(blackList));
-
-        // when & then
-        try {
-            blackListService.addBlackList(userId);
-        } catch (CustomException e) {
-            Assertions.assertThatExceptionOfType(CustomException.class)
-                    .isThrownBy(() -> { throw new CustomException(e.getErrorCode()); })
-                    .withMessageContaining("해당 회원은 이미 블랙입니다.");
-        }
-
-    }
-
-
-    @Test
-    void 블랙리스트에서_삭제한다() {
-
-        // given
-        Long userId = 1L;
-        when(userRepository.findById(userId)).thenReturn(Optional.of(user));
-        when(blackListRepository.findByUserId(user)).thenReturn(Optional.of(blackList));
-
-        // when
-        User result = blackListService.deleteBlackList(userId);
-
-
-        // then
-        Assertions.assertThat(user).isEqualTo(result);
-
-    }
-
-    @Test
-    void 블랙리스트_삭제를_실패한다1() {
-
-        // given
-        Long userId = 1L;
-
-        // when & then
-        try {
-            blackListService.deleteBlackList(userId);
-        } catch (CustomException e) {
-            Assertions.assertThatExceptionOfType(CustomException.class)
-                    .isThrownBy(() -> { throw new CustomException(e.getErrorCode()); })
-                    .withMessageContaining("사용자를 찾을 수 없습니다.");
-        }
-
-    }
-
-    @Test
-    void 블랙리스트_삭제를_실패한다2() {
-
-        // given
-        Long userId = 1L;
-        when(userRepository.findById(userId)).thenReturn(Optional.of(user));
-
-        // when & then
-        try {
-            blackListService.deleteBlackList(userId);
-        } catch (CustomException e) {
-            Assertions.assertThatExceptionOfType(CustomException.class)
-                    .isThrownBy(() -> { throw new CustomException(e.getErrorCode()); })
-                    .withMessageContaining("해당 회원은 블랙이 아닙니다.");
-        }
-
-    }
-
-}
